pyproject.toml
[project]
name = "langchain-1-0-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.3",
    "langchain-google-genai>=3.0.1",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "ruff>=0.14.3",
]
